Phase 2

+++ memtable.h
@@ -8,10 +8,10 @@ using namespace std;
 class Memtable
 {
 public:
-  unsigned int max_size; // Default should be 4KB
-  unsigned int size;     // The number of bytes in the memtable
+  size_t max_size; // Default should be 4KB
+  size_t size;     // The number of bytes in the memtable
 
-  Memtable(uint64_t memtable_size = 0);
+  Memtable(size_t memtable_size = 0);
   db_val_t get(db_key_t key);
   void put(db_key_t key, db_val_t val);
   vector<pair<db_key_t, db_val_t> > scan(db_key_t min_key, db_key_t max_key);
